nothing seems to be committing
Despite multiple attempts to push this through terminal et al, the server keeps timing out. we've tried to get rid of large files; we've tried to do one at a time; we've tried over and over again. but nothing online looks updated. NOT COOL GITHUB!! so trying to upload each project manually as a fork